Disable Blowouts, Anomaly Defs, and Quest Kits
Disabled blowouts for the time being by default.

Redefined a few things for the anomaly artifact spawners to only allow common and uncommon artifacts to avoid doling out 20K+ even on rare occasion.

Moved the Winchester 9422 to Quest Kit 3 to be more in line with what's expected in monetary value from quests.