Various Bugfixes and Updates
Added the option to sell a few things that could eventually be in overabundance in agreement with OpiumCoffee. Made it so Techs have a baseline flag, "2", for access to attachments and upgrades. Prices for these are pending review.

Fixed a major bug with stackable items not updating properly when drag-and-dropped onto one another. Sometimes stacks would be made nil, other times stacks would get all screwed up depending on various circumstances.

Setup new weapons setup with bus flags. Pending review for finalization.