Bugfixes, Weight Update, Tools Update
Updated tech tools to repair for both a ruble cost and a component cost to share the cost and more thoroughly relate the price of the item to the cost to repair it. Cost is quadratic, component cost is linear. Will certainly require balance tweaking, but is good to go functionality wise.