weapon equip selling breaking hud fix
This is a really stupid way of fixing an easy problem. It is 12 am, give me a break plox. When testing on a development server (with a broken version of the gamemode), the client's active weapon would return as the class of the weapon you sold, checking for the item not being in the inv broken the other item. Feel free to revert this commit, it is just a temporary solution for the time being until I or someone finds a better way.