Fixed infinite sprint when out of stamina with CW firearms
A simple check when the 'CW_Move' hook is called to see if the player is out of stamina, forcing them to walk. Before, you would be stopped by the stamina overriding the runspeed but when CW_Move is called after PlayerTick, you would resume running which calls for infinite running/sprint.