Prescription drugs iteminternal not appearing fix
The 'iteminternal' function was not being called when it was ran after the SatisfyAddictions function. This has only occured with drugs prescription drugs. Moving the function above the addiction satisfication function fixes it. Not sure why it is caused, but the common link is that it won't run after 'PrescriptionDrugs' satisfying is ran so I  presume that.