LCD panel task removed
There are 2 tasks that required you to get one or many of an LCD panel which the LCD panels do not spawn automatically as they are not defined in the loot auto spawn table. It has been commented out until the LCD panel is added to the loot auto spawn table or not.