Fix NULL entity weapon raise when dead
If you left click (attack bind) while you are in the DrawDeath menu, it will return a NULL entity error. This is because it is trying to get the players active weapon class which is a nil value because they have no weapon since they are dead. Adding a check for if the player is alive to run the code fixes this.