Fixed card multioptions not registering cards
I committed the isCard check to the item function in an earlier commit but I forgot to commit the part where it registers it inside the individual card deck. On my development server, it was already in the card base, just forgot to commit it to the GitHub... oh well.