PDA money wiring
A system incorporated into the PDA that allows for money wiring - you can pay someone money even if you are not next to them via PDA wiring transactions. It is heavily customisable with the config options. It has been fully tested, including all the config options and anyways to break the system/exploit it. No errors/exploits have been found when testing for them. This took too long to make than it should have.